# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookshelves API",
        "description": "Documentation about Bookshelves API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/authors": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "List of authors",
                "description": "Authors",
                "operationId": "App\\Http\\Controllers\\Api\\AuthorController::index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Integer to choose how many books you show in each page",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 32
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/authors/{author}": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Show author by author slug",
                "description": "Get details for a single author with list of series and books, check /authors endpoint to get list of slugs",
                "operationId": "findAuthorByAuthorSlug",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'auel-jean-m' for Jean M. Auel",
                        "required": true,
                        "style": "form",
                        "example": "auel-jean-m"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid author-slug value or book-slug value",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "List of books",
                "description": "Get list of books with query parameters, default list use pagination.",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::index",
                "parameters": [
                    {
                        "name": "per-page",
                        "in": "query",
                        "description": "Integer to choose how many books you show in each page, default 32",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 32
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Boolean to get list of all books without pagination, default false",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "String to select existant eBook language, default null",
                        "required": false,
                        "schema": {
                            "enum": [
                                "fr",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/books/{author}/{book}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Show book by author slug and book slug",
                "description": "Get details for a single book, check /books endpoint to get list of slugs",
                "operationId": "findBookByAuthorBook",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'lovecraft-howard-phillips' for Howard Phillips Lovecraft",
                        "required": true,
                        "style": "form",
                        "example": "lovecraft-howard-phillips"
                    },
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Slug of book name like 'les-montagnes-hallucinees-fr' for Les Montagnes Hallucinées",
                        "required": true,
                        "style": "form",
                        "example": "les-montagnes-hallucinees-fr"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid author-slug value or book-slug value",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/books/latest": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "List of latest books",
                "description": "Get list of latest books",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::latest",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Integer to choose how many books you show, default 10",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/series/{author}/{series}": {
            "get": {
                "tags": [
                    "series"
                ],
                "summary": "Show series by author slug and by series slug",
                "description": "Get details for a single series with list of books, check /series endpoint to get list of slugs",
                "operationId": "findAuthorByAuthorSlug",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Slug of author name like 'lovecraft-howard-phillips' for Howard Phillips Lovecraft",
                        "required": true,
                        "style": "form",
                        "example": "lovecraft-howard-phillips"
                    },
                    {
                        "name": "series",
                        "in": "path",
                        "description": "Slug of book name like 'cthulhu-le-mythe-fr' for Cthulhu : Le Mythe",
                        "required": true,
                        "style": "form",
                        "example": "cthulhu-le-mythe-fr"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid author-slug value or book-slug value",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}